started the admin backend and added Session and ServiceAvailable models
so far, the admin backend just displays the sales and other models. For
the sales though, the information is divided into two different views.
One is relevant to bakers, and the other is relevant to deliverers. When
the baking is marked as finished, it is removed from the baker's view
and added to the deliverer's view. When the delivery is finished, the
sale is removed from the deliverer's view as well. There is also a
overall Sale view that contains all sales.